+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deprecated
 ### Removed
 ### Fixed
+  * Encode nested array-like payload values as JSON
 ### Security
 
 ## [1.1.0] -- 2017-12-05
@@ -26,4 +27,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.0] -- 2017-11-06
 
 ### Added
-- Initial release
+  * Initial release